Remove SvTAINT from sv_sethek
This was copied from sv_usepvn_flags in commit 58b643af9.  It is
unnecessary, and probably incorrect, as heks are not tainted.

Why sv_sethek used sv_usepvn_flags to begin with I don’t know, but I
imagine it was for brevity’s sake.  This code was ultimately derived
from newSVhek, which doesn’t use sv_usepvn_flags.  Because of that,
and because it is now far enough removed from sv_usepvn_flags, I have
removed the comment referring to it.