bisect-runner.pl should skip the testcase if it fails to build config.sh
Previously if given a testcase and a target of config.sh or config.h, it
would always attempt run it, even if the target (config.sh or config.h)
had not been built. Oddly enough, this usually resulted in the test case
failing, with the consequence that the bisect run ended with a spurious
result. Skipping such revisions instead makes the bisect output meaningful.