Teach bisect-runner.pl how to build a significant PERL_OBJECT commit.
Commit 22c35a8c2392967a is "phase 1 of somewhat major rearrangement of
PERL_OBJECT stuff" but doesn't build due to 2 simple errors. blead in this
broken state was merged to the cfgperl branch, and then these were
immediately corrected there. cfgperl (with the fixes) was merged back to
blead. The resultant rather twisty maze of commits completely confuses git
bisect (and at least me), causing it to the bisect run to confidently return
the wrong answer, an unrelated commit on the cfgperl branch.