Deparse (eof)+1 correctly
If the parentheses are omitted, it means eof(1).

To fix this, we surround the keyword with parentheses (if necessary
according to precedence) if it is an unop, because a final () has a
special meaning for some unops.

For listops, we use surrounding parentheses only if the llafr does
not apply.  Otherwise we use a final ().