Don’t warn for open local *FH
There is this ancient ‘Precedence problem’ warning that warns for open
followed by a bareword and then an operator that would have made open
into a unary-precedence operator in Perl 4.  It was not taking into
account that the bareword might be a Perl keyword.  In that case it
shouldn’t be warning.