Remove explicit support for pre-7.0 VMS and pre-6.0 DEC C.
OpenVMS v7.0 was released in 1995.  There have been no reports of
recent releases of Perl building on versions that far back, yet we
still have quite a bit of code that explicitly supports versions
*prior* to v7.0.

There is a similar story for DEC C v6.0.  It was released in 1998,
and has been superceded by numerous subsequent versions.  Yet the
VMS-specific code in the core is littered with workarounds and
hacks that defend against deficiencies in very old compiler
versions.

This code is for all practical purposes no longer maintained or
maintainable, so the best path forward seems to be to remove it.
Anyone able and willing to commit to long-term support of it
could argue for its restoration, assuming Perl 5.14.x is not
adequate.