utf8.c: Change prototypes of two functions
_to_uni_fold_flags() and _to_fold_latin1() now have their flags
parameter be a boolean.  The name 'flags' is retained in case the usage
ever expands instead of calling it by the name of the only use this
currently has.

This is as a result of confusion between this and _to_ut8_fold_flags()
which does have more than one flag possibility.