Name anon handles __ANONIO__
rather than $__ANONIO__

That dollar sign *has* to have been a mistake.  In ck_fun, the
name was set to __ANONIO__, but it seems the change that added it
(afd1915d43) did not account for the fact that a little later on the
same function checks to makes sure it begins with a dollar sign, as it
could only be a variable name.

rv2gv’s use of $__ANONIO__ (added recently by yours truly) was just
copying was ck_fun was doing.