Add missing <fh> line # phrase to Carp messages
Carp::croak is documented in the camel book as "working like die does,
except misdirecting blame to the caller."

Perldoc says "croak - die of errors (from perspective of caller)".

This is almost true, however careful inspection reveals that croak
does not include file position information ( ", <fh> line $." ) as die
does (when it can).

It turns out that it *is* possible to determine the current file han-
dle (although I wish it were easier).

Carp should do this and add it to the traceback, as it really is use-
ful information.