bisect-runner.pl must not be fooled by a t/test that symlinks to ../miniperl
It had merely been testing that the target file was +x
For t/test it must also check that t/perl wasn't a symlink to miniperl
created as a side effect of make miniperl failing. To add to the fun, for
some revisions the symlink is to '.././perl', not '../perl'.