perlvar: Consistent use of spaces after dots
This commit changes it to use two spaces, for several reasons.  It was
already inconsistent.  Although it used one space in most cases, the
rest of Perl’s documentation uses two.  pod2man expects two spaces, as
it adds two spaces for sentences ending at the end of a line in the
middle of a paragraph, as it was designed to work with Perl’s origi-
nal pod docs, which used two spaces.  If this document used one space
throughout, one would have to go through it and make sure no sentences
end at the ends of lines except at the end of a paragraph.