Teach bisect-runner.pl how to work around various 5.005_5x build problems.
This area of history is particularly hairy because of all the back-and-forth
merging between //depot/perl and //depot/cfgperl, which confuses bisect.
It's also potentially interesting, as it's the start of the 5.6.0 development
series, so many new features arrive in quick succession, making it a likely
spot for regressions to appear.

In particular, we can now build perl for this commit:

  commit 4592e6caefc41a75573f112714d170071892a537
  Author: Nick Ing-Simmons <nik@tiuk.ti.com>
  Date:   Sat May 8 14:16:30 1999 +0000

      Implement OPEN, EOF, SEEK, TELL, BINMODE and FILENO as TIEHANDLE methods.
      Provide Tie::StdHandle
      Basic update of docs.

      p4raw-id: //depot/perl@3330