Update F<Copying> with the FSF's new address.
It is now a byte-for-byte identical copy of the GPL version 1, as downloaded
from http://www.gnu.org/licenses/old-licenses/gpl-1.0.txt

As well as changing their address, at some point the FSF have changed the
amount of indentation on the titles. This commit also reverts our change of
` to '. For this file it seems better to be byte-for-byte identical with the
FSF, even if this does not conform to our formatting.