regcomp.c: regex start class for sharp s
Under most folding types, the optimizer start class should include all
of s, S, and the sharp s (\xdf) if it includes any of them.  The code
was neglecting the latter.  This is currently not relevant, as there is
special handling of the sharp s elsewhere in regcomp.c.  But this is a
step to changing that special handling to fix some bugs.