regcomp.c: Need to account for delta sizes
When a node can match varying sizes, the delta variable in the optimizer
needs to change to account for that, and it can no longer match a fixed
length string.

This code was adapted from the existing code for the FOLDCHAR node that
has to deal with the same problem.