regex: Add new node type EXACTFU_SS
This node will be used to distinguish between the case in a non-UTF8
pattern and string where something could be matched that is of different
lengths.  The only instance where this can happen is the LATIN SMALL
LETTER SHARP S can match the sequences "ss", "Ss", "sS", or "SS", hence
the name.

This node is not currently generated; this prepares for future commits