regex: Remove FOLDCHAR regnode type
This node type hasn't been used since 5.14.0.  Instead an ANYOFV node
was generated where formerly a FOLDCHAR node would have been used.  The
ANYOFV was used because it already existed and was up-to-date, whereas
FOLDCHAR would have needed some bug fixes to adapt it, even though it
would be faster in execution than ANYOFV; so the code for it was
retained in case it was needed.

However, both these solutions were defective, and a previous commit has
changed things to a different type of solution entirely.  Thus FOLDCHAR
is obsolescent and can be removed, though the code in it was used as a
base for some of the new solutions.