re/reg_fold.t: Test more code points
This statement was wrong that said all these things are tested in
fold_grind.t.  It will test them all when run with a particular option,
but due to time issues, it skips many code points.  reg_fold.t, on the
other hand, does just basic sanity testing, and so should always test
every code point for that.