mktables: Account for 2 provisional files
This is in preparation for Unicode 6.1.  These two provisional
properties are available starting in 6.0, but the rest of the data base
did not refer to them, so we could safely ignore them.  But in the beta
6.1, they are referred to elsewhere, which makes mktables generate hooks
for them, and then complains when those hooks are unused.

This patch should work no matter how 6.1 final comes out, and allows
someone who really wants to use these provisional properties to
recompile Perl without having to delve too far into the mktables code to
get it to work.